refactor(web): decompose PlaygroundPage into 6 colocated components (#453)
PlaygroundPage went from 813 lines down to 518 (−36%). Six new components under 'components/playground/':

- PlaygroundHelpers (101L) — pure helpers (extractEnvVarKeys, isRuntimeBased, defaultPromptStarters) + ThinkingBubble pre-token streaming indicator
- PlaygroundEmptyHero (86L) — ChatGPT-style centered welcome flag with skill identity, lead sentence, 3 suggestion chips, and a drawer-discovery footer. Disabled state on the chips mirrors the chat-input lock.
- PlaygroundConversation (109L) — forwardRef component: rendered turns + streaming buffer + ThinkingBubble + file outputs (image inline, others as download link) + error banner + the bottom anchor div the parent scrolls into view.
- PlaygroundRail (101L) — fixed right-edge vertical rail with hover-peek + click-to-pin + left-side tooltip + warning dot + active-pin marker. Stateless; parent owns the state.
- PlaygroundEnvDrawerBody (81L) — per-key input list with Set/Required pill + lock hint when any required key is empty.
- PlaygroundPackageDrawerBody (79L) — SkillPackagePreview wrapper with a registry-link footer.

What's still in PlaygroundPage: state plumbing (hooks for skill / package / chat / quota), refs (chat input handle, scroll anchor, scroll-stick tracker), the drawer outer container (motion.aside + backdrop), the composer (ChatInput + ModelPicker + QuotaInline), and the early-return states (no-skill / loading / 404 / over-limit).

The drawer-outer motion container is hard to bisect cleanly: it owns the rail-tab close timer, the AnimatePresence boundary, the pin/unpin button, and the [§ NAME] header — too many narrow props to pull out without making the contract uglier than the inline version.

Issue #453 stays open until the suggested 'usePlaygroundSession()' hook extraction goes in — that would pull queries / chat state / handlers out of the page entirely and likely bring it under 300L. Deferred because the data flow weaves through the JSX layout in ways that don't bisect into a small commit.

Per the issue's 'one PR per page' guidance: SkillDetailPage #651, DocsPage #659, this PR closes the page-component portion of #453.

Behavior unchanged; 110 web tests still pass.
